Small refacto
