+++ src/Constants.ts
@@ -1,22 +1,55 @@
+// Canvas constants
+const REFRESH_INTERVAL = 30
 const CANVAS_WIDTH = 1200
 const CANVAS_HEIGHT = 720
 const MENU_UPGRADE_WIDTH = 300
 const MENU_UPGRADE_MARGIN = 50
 
+// Game constants
+const NUMBER_OF_LIVES = 3
 const INVINCIBILITY_TIME = 3
 const LASER_SHOOTING_RATE = 2
 const LASER_RANGE = 500
+
+// Ship constants
+const SHIP_WIDTH = 10
+const SHIP_LENGTH = 20
 const SHIP_MAX_SPEED = 3
 const SHIP_ACCELERATION = 3
 
+// Asteroids constants
+const ASTEROID_MAX_SPEED = 100
+const ASTEROID_SIZE = 20
+const GIGA_ASTEROID_SIZE = 200
+const ASTEROID_JAGGEDNESS = 0.3
+const ASTEROID_PARTICULE_LENGTH = 2
+const ASTEROID_PARTICULE_MIN_SPEED = 50
+const ASTEROID_PARTICULE_MAX_SPEED = 200
+const ASTEROID_PARTICULE_MAX_RANGE = 100
+
 export {
+    REFRESH_INTERVAL,
     CANVAS_WIDTH,
     CANVAS_HEIGHT,
     MENU_UPGRADE_WIDTH,
     MENU_UPGRADE_MARGIN,
+    NUMBER_OF_LIVES,
     INVINCIBILITY_TIME,
     LASER_SHOOTING_RATE,
     LASER_RANGE,
+    SHIP_WIDTH,
+    SHIP_LENGTH,
     SHIP_MAX_SPEED,
     SHIP_ACCELERATION,
+}
+
+export {
+    ASTEROID_MAX_SPEED,
+    ASTEROID_SIZE,
+    GIGA_ASTEROID_SIZE,
+    ASTEROID_JAGGEDNESS,
+    ASTEROID_PARTICULE_LENGTH,
+    ASTEROID_PARTICULE_MIN_SPEED,
+    ASTEROID_PARTICULE_MAX_SPEED,
+    ASTEROID_PARTICULE_MAX_RANGE,
 }
